Split GenericDictionaryAssertionSpecs (#2424)
